Safari-609.1.11.1/LayoutTests/fast/animation/request-animation-frame-throttle-inside-overflow-scroll-expected.txt

Tests that requestAnimationFrame is throttled in subframes are inside scrolled overflow scroll

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Frame is initially inside the viewport so requestAnimationFrame should not be throttled
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() is false
Scrolling overflow.
requestAnimationFrame should still not be throttled
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() is false
PASS successfullyParsed is true

TEST COMPLETE
